requirement.txt
requests
sqlite3
win32crypt
cryptography
selenium